Moved LetterboxingViewport from app to graphics. #183
